+++ kiwi/boot/image/dracut.py
@@ -16,8 +16,6 @@
 # along with kiwi.  If not, see <http://www.gnu.org/licenses/>
 #
 import os
-import re
-from collections import namedtuple
 
 # project
 from kiwi.logger import log
@@ -27,9 +25,6 @@ from kiwi.boot.image.base import BootImageBase
 from kiwi.defaults import Defaults
 from kiwi.system.profile import Profile
 from kiwi.system.setup import SystemSetup
-from kiwi.path import Path
-
-from kiwi.exceptions import KiwiDiskBootImageError
 
 
 class BootImageDracut(BootImageBase):
@@ -202,68 +197,3 @@ class BootImageDracut(BootImageBase):
             self.initrd_filename = os.sep.join(
                 [self.target_dir, dracut_initrd_basename]
             )
-
-    def get_boot_names(self):
-        """
-        Provides kernel and initrd names for kiwi boot image
-
-        :return:
-            Contains boot_names_type tuple
-
-            .. code:: python
-
-                boot_names_type(
-                    kernel_name='INSTALLED_KERNEL',
-                    initrd_name='DRACUT_OUTPUT_NAME'
-                )
-
-        :rtype: tuple
-        """
-        boot_names_type = namedtuple(
-            'boot_names_type', ['kernel_name', 'initrd_name']
-        )
-        kernel = Kernel(
-            self.boot_root_directory
-        )
-        kernel_info = kernel.get_kernel()
-        if not kernel_info:
-            raise KiwiDiskBootImageError(
-                'No kernel in boot image tree %s found' %
-                self.boot_root_directory
-            )
-        dracut_output_format = self._get_dracut_output_file_format()
-        return boot_names_type(
-            kernel_name=kernel_info.name,
-            initrd_name=dracut_output_format.format(
-                kernel_version=kernel_info.version
-            )
-        )
-
-    def _get_dracut_output_file_format(self):
-        """
-        Unfortunately the dracut initrd output file format varies between
-        the different Linux distributions. Tools like lsinitrd, and also
-        grub2 rely on the initrd output file to be in that format.
-        Thus when kiwi uses dracut the same file format should be used
-        all over the place in order to stay compatible with what the
-        distribution does
-        """
-        default_outfile_format = 'initramfs-{kernel_version}.img'
-        dracut_search_env = {
-            'PATH': os.sep.join([self.boot_root_directory, 'usr', 'bin'])
-        }
-        dracut_tool = Path.which(
-            'dracut', custom_env=dracut_search_env, access_mode=os.X_OK
-        )
-        if dracut_tool:
-            outfile_expression = r'outfile="/boot/(init.*\$kernel.*)"'
-            with open(dracut_tool) as dracut:
-                matches = re.findall(outfile_expression, dracut.read())
-                if matches:
-                    return matches[0].replace('$kernel', '{kernel_version}')
-
-        log.warning('Could not detect dracut output file format')
-        log.warning('Using default initrd file name format {0}'.format(
-            default_outfile_format
-        ))
-        return default_outfile_format
