+++ test/unit/boot_image_builtin_kiwi_test.py
@@ -2,7 +2,6 @@ import sys
 import mock
 from mock import patch
 from mock import call
-from collections import namedtuple
 from pytest import raises
 
 import kiwi
@@ -165,14 +164,6 @@ class TestBootImageKiwi:
         mock_path.return_value = True
         self.boot_image.__del__()
         mock_wipe.assert_called_once_with('boot-root-directory')
-
-    def test_boot_names(self):
-        boot_names_type = namedtuple(
-            'boot_names_type', ['kernel_name', 'initrd_name']
-        )
-        assert self.boot_image.get_boot_names() == boot_names_type(
-            kernel_name='linux.vmx', initrd_name='initrd.vmx'
-        )
 
     def teardown(self):
         sys.argv = argv_kiwi_tests
